mostly fix shaderpackoptionlist clicking buttons

+++ src/main/java/net/coderbot/iris/gui/element/widget/AbstractElementWidget.java
@@ -20,11 +20,11 @@ public abstract class AbstractElementWidget<T extends OptionMenuElement> {
 
 	public abstract void render(int x, int y, int width, int height, int mouseX, int mouseY, float tickDelta, boolean hovered);
 
-	public boolean mouseClicked(double mx, double my, int button) {
+	public boolean mouseClicked(int mx, int my, int button) {
 		return false;
 	}
 
-	public boolean mouseReleased(double mx, double my, int button) {
+	public boolean mouseReleased(int mx, int my, int button) {
 		return false;
 	}
 }
